File structure organized, css files created and proptypes added
