+++ src/components/Footer/Footer.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import './Footer.css'
 
 const Footer = () => {
   return <div>Footer</div>
