+++ src/components/Header/Header.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import logo from '../../logo.svg'
+import './Header.css'
 
 const Header = () => {
   return (
