+++ src/components/Layout/Layout.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import Header from '../Header/Header'
-import Footer from '../Footer/Footer'
+import Header from '../../components/Header/Header'
+import Footer from '../../components/Footer/Footer'
+import './Layout.css'
 
 const Layout = ({ children }) => {
   return (
