+++ src/containers/Restaurants/Restaurants.js
@@ -2,10 +2,11 @@ import React, { useEffect, useState } from 'react'
 import { uid } from 'uid'
 import { useDispatch, useSelector } from 'react-redux'
 import { getRestaurants } from '../../actions/restaurants'
-import Restaurant from '../Restaurant/Restaurant'
-import LoadingSpinner from '../LoadingSpinner/LoadingSpinner'
+import Restaurant from '../../components/Restaurant/Restaurant'
+import LoadingSpinner from '../../components/LoadingSpinner/LoadingSpinner'
 import Pagination from '../Pagination/Pagination'
 import { setCurrentPageNumber } from '../../actions/session'
+import './Restaurants.css'
 
 const Restaurants = () => {
   const dispatch = useDispatch()
