+++ src/pages/RestaurantsPage/RestaurantsPage.js
@@ -1,0 +1,14 @@
+import React from 'react'
+import Restaurants from '../../containers/Restaurants/Restaurants'
+import Form from '../../containers/Form/Form'
+
+const RestaurantsPage = () => {
+  return (
+    <section>
+      <Form />
+      <Restaurants />
+    </section>
+  )
+}
+
+export default RestaurantsPage
